BZ-1344 support multi join
