+++ scala-client/src/main/java/tech/mlsql/byzer_client_sdk/scala_lang/generator/byzer.scala
@@ -1,6 +1,7 @@
 package tech.mlsql.byzer_client_sdk.scala_lang.generator
 
 import net.sf.json.{JSONArray, JSONObject}
+import tech.mlsql.byzer_client_sdk.scala_lang.generator.node.{Columns, ET, Filter, Include, Join, Load, OrderBy, Python, Raw, Register, Save, Set}
 import tech.mlsql.common.utils.serder.json.JsonUtils
 
 import scala.collection.JavaConverters._
