+++ scala-client/src/test/java/tech/mlsql/test/byzer_client_sdk/scala_lang/generator/ByzerScriptTest.scala
@@ -2,6 +2,7 @@ package tech.mlsql.test.byzer_client_sdk.scala_lang.generator
 
 import org.scalatest.funsuite.AnyFunSuite
 import tech.mlsql.byzer_client_sdk.scala_lang.generator._
+import tech.mlsql.byzer_client_sdk.scala_lang.generator.node.{Columns, Filter, JoinMeta}
 import tech.mlsql.common.utils.serder.json.JSONTool
 
 /**
@@ -94,26 +95,112 @@ class ByzerScriptTest extends AnyFunSuite {
     println(genCode)
   }
 
+  test("join-multi") {
+    val byzer = Byzer()
+    val table1 = byzer.load.format("csv").path("/tmp/jack").options().add("header", "true").end
+    val table2 = byzer.load.format("csv").path("/tmp/william").options().add("header", "true").end
+    val table3 = byzer.load.format("csv").path("/tmp/admond").options().add("header", "true").end
+    table1.namedTableName("table1").end
+    table2.namedTableName("table2").end
+    table3.namedTableName("table3").end
+
+    val t1 = table1.tableName
+    val t2 = table2.tableName
+    val t3 = table3.tableName
+
+    val genCode = byzer.join.
+      from(Expr(Some(t1))).
+      leftColumns(Expr(Some(s"${t1}.a,${t1}.c"))).
+      rightColumns(Expr(Some(s"""${t2}.m,${t3}.n"""))).
+      rightColumns(Expr(Some(s"""${t2}.g,${t3}.t"""))).
+      left(Expr(Some(t2))).
+      on_and.add(Expr(Some(s"""${t1}.a=${t2}.b"""))).add(Expr(Some(s"""${t1}.a=${t2}.c"""))).end.
+      right(Expr(Some(t3))).
+      on_or.add(Expr(Some(s"""${t1}.a=${t3}.d"""))).add(Expr(Some(s"""${t1}.a=${t3}.e"""))).add(Expr(Some(s"""${t1}.a=${t3}.f"""))).end.
+      namedTableName("outputTable").end.toScript
+
+    val target =
+      """load csv.`/tmp/jack` where `header`='''true''' as table1;
+        |load csv.`/tmp/william` where `header`='''true''' as table2;
+        |load csv.`/tmp/admond` where `header`='''true''' as table3;
+        |select table1.a,table1.c,table2.m,table3.n,table2.g,table3.t
+        |from table1
+        |LEFT OUTER JOIN table2 on (table1.a=table2.b and table1.a=table2.c)
+        |RIGHT OUTER JOIN table3 on ((table1.a=table3.d or table1.a=table3.e) or table1.a=table3.f)
+        |as outputTable;""".stripMargin
+    assert(genCode == target, "Generate multi join clause error!")
+  }
+
+  test("join-multi-error") {
+    val byzer = Byzer()
+    val table1 = byzer.load.format("csv").path("/tmp/jack").options().add("header", "true").end
+    val table2 = byzer.load.format("csv").path("/tmp/william").options().add("header", "true").end
+    val table3 = byzer.load.format("csv").path("/tmp/admond").options().add("header", "true").end
+    table1.namedTableName("table1").end
+    table2.namedTableName("table2").end
+    table3.namedTableName("table3").end
+
+    val t1 = table1.tableName
+    val t2 = table2.tableName
+    val t3 = table3.tableName
+
+    val caught =
+      intercept[RuntimeException] {
+        byzer.join.
+          from(Expr(Some(t1))).
+          leftColumns(Expr(Some(s"${t1}.a,${t1}.c"))).
+          rightColumns(Expr(Some(s"""${t2}.m,${t3}.n"""))).
+          rightColumns(Expr(Some(s"""${t2}.g,${t3}.t"""))).
+          left(Expr(Some(t2))).
+          on_and.add(Expr(Some(s"""${t1}.a=${t2}.b"""))).add(Expr(Some(s"""${t1}.a=${t2}.c"""))).end.
+          right(Expr(Some(t3))).
+          on_or.add(Expr(Some(s"""${t1}.a=${t3}.d"""))).add(Expr(Some(s"""${t1}.a=${t3}.e"""))).add(Expr(Some(s"""${t1}.a=${t3}.f"""))).end.
+          left(Expr(Some(t2))).
+          namedTableName("outputTable").end.toScript
+      }
+    assert(caught.getMessage == "3 join() but 2 on(), the join method must correspond to the on method one by one!")
+  }
+
+
   test("join-serder") {
     val byzer = Byzer()
     val table1 = byzer.load.format("csv").path("/tmp/jack").options().add("header", "true").end
     val table2 = byzer.load.format("csv").path("/tmp/william").options().add("header", "true").end
-    table1.end
-    table2.end
+    val table3 = byzer.load.format("csv").path("/tmp/admond").options().add("header", "true").end
+    table1.namedTableName("table1").end
+    table2.namedTableName("table2").end
+    table3.namedTableName("table3").end
 
     val t1 = table1.tableName
     val t2 = table2.tableName
+    val t3 = table3.tableName
 
     val f = byzer.join.
       from(Expr(Some(t1))).
       left(Expr(Some(t2))).
       on_and.add(Expr(Some(s"""${t1}.a=${t2}.b"""))).add(Expr(Some(s"""${t1}.a=${t2}.c"""))).end.
       leftColumns(Expr(Some(s"${t1}.a,${t1}.c"))).
-      rightColumns(Expr(Some(s"""${t2}.m"""))).end
-    println(f.toJson(true))
-    val v = Byzer().fromJson(f.toJson())
+      rightColumns(Expr(Some(s"""${t2}.m"""))).
+      right(Expr(Some(t3))).
+      on_and.add(Expr(Some(s"""${t1}.a=${t3}.d"""))).add(Expr(Some(s"""${t1}.a=${t3}.e"""))).end.
+      rightColumns(Expr(Some(s"""${t3}.n""")))
+      .namedTableName("joinTable").end
 
-    println(v.toScript)
+    val jsonStr = f.toJson(true)
+    println(jsonStr)
+
+    val byzerV2 = Byzer().fromJson(f.toJson())
+
+    val script = byzerV2.toScript
+    val expectScript = """load csv.`/tmp/jack` where `header`='''true''' as table1;
+                         |load csv.`/tmp/william` where `header`='''true''' as table2;
+                         |load csv.`/tmp/admond` where `header`='''true''' as table3;
+                         |select table1.a,table1.c,LEFT OUTER JOIN table2,RIGHT OUTER JOIN table3
+                         |from table1
+                         |LEFT OUTER JOIN table2 on (table1.a=table2.b and table1.a=table2.c)
+                         |RIGHT OUTER JOIN table3 on (table1.a=table3.d and table1.a=table3.e)
+                         |as joinTable;""".stripMargin
+    assert(script == expectScript)
   }
 
   /**
